Update README, fix some testing suite
The README.md now should be accurate again for setting up a vagrant instance (it's a little more involved than it used to be.

Also got the rake test suite working again.
The trick was just switching to an SQL schema file, as rake was dropping and recreating the test database every time, and the rails schema didn't explain the view.
The FURTHER trick was commenting out a dubious line in that schema, which I assume I'll have to do every time, which is a little annoying.